Extract the skill value unions into a leaf types module
skill_configuration.ts imports MCPServerViewSchema as a runtime value, which
drags in mcp_internal_actions/constants.ts and therefore every MCP server's
metadata — so a server metadata file importing it closes a cycle and throws at
module init. Move SKILL_STATUSES/SKILL_AVAILABILITIES to a dependency-free
module next to it, re-export them so existing import paths keep working, and
drop the copies workspace_management had to inline.
